+++ internal/dao/dao.go
@@ -32,17 +32,17 @@ type Daos interface {
 	CreateUser(ctx context.Context, user *model.User) error
 
 	SaveAnswerSheet(ctx context.Context, answerSheet AnswerSheet) error
-	GetAnswerSheetBySurveyID(ctx context.Context, surveyID int, pageNum int, pageSize int) (
+	GetAnswerSheetBySurveyID(ctx context.Context, surveyID int64, pageNum int, pageSize int) (
 		[]AnswerSheet, *int64, error)
-	DeleteAnswerSheetBySurveyID(ctx context.Context, surveyID int) error
+	DeleteAnswerSheetBySurveyID(ctx context.Context, surveyID int64) error
 	DeleteAnswerSheetByAnswerID(ctx context.Context, AnswerID primitive.ObjectID) error
 	GetAnswerSheetByAnswerID(ctx context.Context, AnswerID primitive.ObjectID) (AnswerSheet, error)
 
-	CreateManage(ctx context.Context, id int, surveyID int) error
-	DeleteManage(ctx context.Context, id int, surveyID int) error
-	DeleteManageBySurveyID(ctx context.Context, surveyID int) error
-	CheckManage(ctx context.Context, id int, surveyID int) error
-	GetManageByUIDAndSID(ctx context.Context, uid int, sid int) (*model.Manage, error)
+	CreateManage(ctx context.Context, id int, surveyID int64) error
+	DeleteManage(ctx context.Context, id int, surveyID int64) error
+	DeleteManageBySurveyID(ctx context.Context, surveyID int64) error
+	CheckManage(ctx context.Context, id int, surveyID int64) error
+	GetManageByUIDAndSID(ctx context.Context, uid int, sid int64) (*model.Manage, error)
 	GetManageByUserID(ctx context.Context, uid int) ([]model.Manage, error)
 
 	CreateOption(ctx context.Context, option *model.Option) error
@@ -50,20 +50,20 @@ type Daos interface {
 	DeleteOption(ctx context.Context, optionID int) error
 
 	CreateQuestion(ctx context.Context, question *model.Question) error
-	GetQuestionsBySurveyID(ctx context.Context, surveyID int) ([]model.Question, error)
+	GetQuestionsBySurveyID(ctx context.Context, surveyID int64) ([]model.Question, error)
 	GetQuestionByID(ctx context.Context, questionID int) (*model.Question, error)
 	DeleteQuestion(ctx context.Context, questionID int) error
-	DeleteQuestionBySurveyID(ctx context.Context, surveyID int) error
+	DeleteQuestionBySurveyID(ctx context.Context, surveyID int64) error
 
 	CreateSurvey(ctx context.Context, survey *model.Survey) error
-	GetSurveyByID(ctx context.Context, surveyID int) (*model.Survey, error)
+	GetSurveyByID(ctx context.Context, surveyID int64) (*model.Survey, error)
 	GetSurveyByTitle(ctx context.Context, title string, num, size int) ([]model.Survey, *int64, error)
-	DeleteSurvey(ctx context.Context, surveyID int) error
-	UpdateSurveyStatus(ctx context.Context, surveyID int, status int) error
-	UpdateSurvey(ctx context.Context, id int, title, desc, img string, deadline time.Time) error
+	DeleteSurvey(ctx context.Context, surveyID int64) error
+	UpdateSurveyStatus(ctx context.Context, surveyID int64, status int) error
+	UpdateSurvey(ctx context.Context, id int64, title, desc, img string, deadline time.Time) error
 	GetAllSurveyByUserID(ctx context.Context, userId int) ([]model.Survey, error)
-	IncreaseSurveyNum(ctx context.Context, sid int) error
+	IncreaseSurveyNum(ctx context.Context, sid int64) error
 
-	SaveRecordSheet(ctx context.Context, answerSheet RecordSheet, sid int) error
-	DeleteRecordSheets(ctx context.Context, surveyID int) error
+	SaveRecordSheet(ctx context.Context, answerSheet RecordSheet, sid int64) error
+	DeleteRecordSheets(ctx context.Context, surveyID int64) error
 }
