+++ internal/dao/manage.go
@@ -7,32 +7,32 @@ import (
 )
 
 // CreateManage 创建问卷权限
-func (d *Dao) CreateManage(ctx context.Context, id int, surveyID int) error {
+func (d *Dao) CreateManage(ctx context.Context, id int, surveyID int64) error {
 	err := d.orm.WithContext(ctx).Create(&model.Manage{UserID: id, SurveyID: surveyID}).Error
 	return err
 }
 
 // DeleteManage 删除问卷权限
-func (d *Dao) DeleteManage(ctx context.Context, id int, surveyID int) error {
+func (d *Dao) DeleteManage(ctx context.Context, id int, surveyID int64) error {
 	err := d.orm.WithContext(ctx).Where("user_id = ? AND survey_id = ?", id, surveyID).Delete(&model.Manage{}).Error
 	return err
 }
 
 // DeleteManageBySurveyID 根据问卷ID删除问卷权限
-func (d *Dao) DeleteManageBySurveyID(ctx context.Context, surveyID int) error {
+func (d *Dao) DeleteManageBySurveyID(ctx context.Context, surveyID int64) error {
 	err := d.orm.WithContext(ctx).Where("survey_id = ?", surveyID).Delete(&model.Manage{}).Error
 	return err
 }
 
 // CheckManage 检查问卷权限
-func (d *Dao) CheckManage(ctx context.Context, id int, surveyID int) error {
+func (d *Dao) CheckManage(ctx context.Context, id int, surveyID int64) error {
 	var manage model.Manage
 	err := d.orm.WithContext(ctx).Where("user_id = ? AND survey_id = ?", id, surveyID).First(&manage).Error
 	return err
 }
 
 // GetManageByUIDAndSID 根据用户ID和问卷ID获取问卷权限
-func (d *Dao) GetManageByUIDAndSID(ctx context.Context, uid int, sid int) (*model.Manage, error) {
+func (d *Dao) GetManageByUIDAndSID(ctx context.Context, uid int, sid int64) (*model.Manage, error) {
 	var manage model.Manage
 	err := d.orm.WithContext(ctx).Where("user_id = ? AND survey_id = ?", uid, sid).First(&manage).Error
 	return &manage, err
