+++ internal/dao/survey.go
@@ -16,13 +16,13 @@ func (d *Dao) CreateSurvey(ctx context.Context, survey model.Survey) (model.Surv
 }
 
 // UpdateSurveyStatus 更新问卷状态
-func (d *Dao) UpdateSurveyStatus(ctx context.Context, surveyID int, status int) error {
+func (d *Dao) UpdateSurveyStatus(ctx context.Context, surveyID int64, status int) error {
 	err := d.orm.WithContext(ctx).Model(&model.Survey{}).Where("id = ?", surveyID).Update("status", status).Error
 	return err
 }
 
 // UpdateSurvey 更新问卷
-func (d *Dao) UpdateSurvey(ctx context.Context, id int, surveyType, limit uint,
+func (d *Dao) UpdateSurvey(ctx context.Context, id int64, surveyType, limit uint,
 	sumLimit uint, verify bool, desc string, title string, deadline, startTime time.Time) error {
 	err := d.orm.WithContext(ctx).Model(&model.Survey{}).Where("id = ?", id).
 		Updates(model.Survey{
@@ -46,7 +46,7 @@ func (d *Dao) GetSurveyByUserID(ctx context.Context, userId int) ([]model.Survey
 }
 
 // GetSurveyByID 根据问卷ID获取问卷
-func (d *Dao) GetSurveyByID(ctx context.Context, surveyID int) (*model.Survey, error) {
+func (d *Dao) GetSurveyByID(ctx context.Context, surveyID int64) (*model.Survey, error) {
 	var survey model.Survey
 	err := d.orm.WithContext(ctx).Where("id = ?", surveyID).First(&survey).Error
 	return &survey, err
@@ -60,14 +60,14 @@ func (d *Dao) GetAllSurvey(ctx context.Context) ([]model.Survey, error) {
 }
 
 // IncreaseSurveyNum 增加问卷填写人数
-func (d *Dao) IncreaseSurveyNum(ctx context.Context, sid int) error {
+func (d *Dao) IncreaseSurveyNum(ctx context.Context, sid int64) error {
 	err := d.orm.WithContext(ctx).Model(&model.Survey{}).Where("id = ?", sid).
 		Update("num", gorm.Expr("num + ?", 1)).Error
 	return err
 }
 
 // DeleteSurvey 删除问卷
-func (d *Dao) DeleteSurvey(ctx context.Context, surveyID int) error {
+func (d *Dao) DeleteSurvey(ctx context.Context, surveyID int64) error {
 	err := d.orm.WithContext(ctx).Where("id = ?", surveyID).Delete(&model.Survey{}).Error
 	return err
 }
